refactor: export client, hooks and modelsfrom api folder
